requirements.txt
uvicorn>=0.24.0,<0.30.0
fastapi~=0.104.1
python-multipart>=0.0.5,<0.1.0
pydantic~=2.4.2
typing_extensions>=4.2.0,<5.0.0
loguru>=0.7.2,<1.0.0
pydantic_settings~=2.0.3
git+https://github.com/baremetals/careerappmodelpackage.git#egg=career_app_model
numpy~=1.26.1
joblib~=1.3.2
pandas~=2.1.2
pymilvus~=2.3.2
psycopg2-binary~=2.9.9
sqlalchemy~=2.0.23
alembic~=1.13.0
passlib~=1.7.4
bcrypt~=4.1.1
starlette~=0.27.0
python-jose~=3.3.0